Add initial pidfile support for #21
So on the plus side this is doable and was pretty easy to implement.
Downsides are:

1. I have to bind to the internals of System.Process. This breaks
windows compatibility, which is fine because this package already uses
the `unix` package.
2. I will need to be fairly strict about the version of the Process
library. While the logic of managing processes in general won't change
any time soon, the way they internally represent processes might.

Still need to:

1. bump version.
2. release to hackage given Jamie's goahead if it is possible for me to
obtain it.